resources/requirements.txt
langchain_community
langchain_experimental
dotenv
google.generativeai
psycopg2-binary